+++ packages/react/tests/e2e-storybook/cypress/support/index.js
@@ -18,7 +18,7 @@ beforeEach(() => {
   // Mock the translation file
   cy.intercept(
     'https://1.www.s81c.com/common/carbon-for-ibm-dotcom/translations/masthead-footer/usen.json',
-    { fixture: 'translation.json' }
+    { fixture: 'translation-raw.json' }
   );
 
   // Mock the user status