+++ packages/web-components/tests/e2e-storybook/cypress/integration/masthead/masthead-custom-search.e2e.js
@@ -17,6 +17,8 @@ describe('dds-masthead | custom search (desktop)', () => {
   beforeEach(() => {
     cy.visit(`/${_pathCustomSearch}`);
     cy.viewport(1280, 780);
+
+    cy.waitUntil(() => cy.get('[data-autoid="dds--masthead-default__l0-nav0"]').should('not.be.empty'));
   });
 
   it('should open the search bar on click', () => {
@@ -29,6 +31,11 @@ describe('dds-masthead | custom search (desktop)', () => {
   });
 
   it('should display grouped results with hrefs', () => {
+    // Mock grouped search typeahead API
+    cy.intercept('https://ibmdocs-dev.mybluemix.net/docs/api/v1/suggest?query=cloud&categories=&limit=6', {
+      fixture: 'grouped-typeahead.json',
+    });
+
     cy.get('dds-masthead > dds-search-with-typeahead')
       .shadow()
       .find('.bx--header__search--search')