+++ packages/web-components/tests/e2e-storybook/cypress/integration/masthead/masthead-scoped-search.e2e.js
@@ -13,10 +13,12 @@
  */
 const _pathScopedSearch = '/iframe.html?id=components-masthead--with-scoped-search';
 
-xdescribe('dds-masthead | scoped search (desktop)', () => {
+describe('dds-masthead | scoped search (desktop)', () => {
   beforeEach(() => {
     cy.visit(`/${_pathScopedSearch}`);
     cy.viewport(1280, 780);
+
+    cy.waitUntil(() => cy.get('[data-autoid="dds--masthead-default__l0-nav0"]').should('not.be.empty'));
   });
 
   it('should open the search bar on click', () => {
@@ -47,6 +49,11 @@ xdescribe('dds-masthead | scoped search (desktop)', () => {
   });
 
   it('should retrieve less results with "pw" scope', () => {
+    // Mock scoped search typeahead API
+    cy.intercept('**/search/typeahead/v1?*', {
+      fixture: 'scoped-typeahead.json',
+    });
+
     cy.get('dds-masthead > dds-search-with-typeahead')
       .shadow()
       .find('.bx--header__search--search')
@@ -73,10 +80,12 @@ xdescribe('dds-masthead | scoped search (desktop)', () => {
   });
 });
 
-xdescribe('dds-masthead | scoped search (mobile)', () => {
+describe('dds-masthead | scoped search (mobile)', () => {
   beforeEach(() => {
     cy.visit(`/${_pathScopedSearch}`);
     cy.viewport(320, 780);
+
+    cy.waitUntil(() => cy.get('[data-autoid="dds--masthead-default__l0-nav0"]').should('not.be.empty'));
   });
 
   it('should open the search bar on click', () => {
@@ -107,6 +116,11 @@ xdescribe('dds-masthead | scoped search (mobile)', () => {
   });
 
   it('should retrieve less results with "pw" scope', () => {
+    // Mock scoped search typeahead API
+    cy.intercept('**/search/typeahead/v1?*', {
+      fixture: 'scoped-typeahead.json',
+    });
+
     cy.get('dds-masthead > dds-search-with-typeahead')
       .shadow()
       .find('.bx--header__search--search')